test(davinci): satisfy helper order gates
