+++ crates/vize_s1_to_s2/src/emit/buf.rs
@@ -13,11 +13,8 @@ pub(super) struct Buf {
     used: u64,
     /// Transform-analogue registration order (`root.helpers`).
     preferred: StdVec<Helper>,
-    /// First `use_*` order (`used_helpers`). Same-rank leftovers usually
-    /// follow this, not `Helper::ALL`, so a builtin tag used before `withCtx`
-    /// stays before it (`Transition` then `withCtx`; nested `KeepAlive`
-    /// after the parent's `withCtx`). Modifier and normalize helpers are
-    /// ordered by final alias use to match shipped output.
+    /// First `use_*` order (`used_helpers`), with modifier and normalize
+    /// helpers reordered by final alias use to match shipped output.
     used_order: StdVec<Helper>,
     /// Compact static-props / vnode RHS values, `_hoisted_1` first.
     hoists: StdVec<String>,
@@ -379,47 +376,4 @@ impl Buf {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::{Buf, Helper};
-
-    #[test]
-    fn helper_preamble_uses_final_body_order_within_one_rank() {
-        let mut buf = Buf::new();
-        buf.use_helper(Helper::NormalizeStyle);
-        buf.use_helper(Helper::NormalizeClass);
-        buf.push("_normalizeClass(cls); _normalizeStyle(style)");
-
-        assert!(buf.preamble().starts_with(
-            "const { normalizeClass: _normalizeClass, normalizeStyle: _normalizeStyle"
-        ));
-    }
-
-    #[test]
-    fn helper_preamble_uses_final_hoist_order_within_one_rank() {
-        let mut buf = Buf::new();
-        buf.use_helper(Helper::NormalizeStyle);
-        buf.use_helper(Helper::NormalizeClass);
-        buf.push_hoist("_normalizeClass(cls)".into());
-        buf.push_hoist("_normalizeStyle(style)".into());
-
-        assert!(buf.preamble().starts_with(
-            "const { normalizeClass: _normalizeClass, normalizeStyle: _normalizeStyle"
-        ));
-    }
-
-    #[test]
-    fn helper_preamble_keeps_preferred_before_body_order() {
-        let mut buf = Buf::new();
-        buf.prefer(Helper::WithDirectives);
-        buf.use_helper(Helper::WithKeys);
-        buf.use_helper(Helper::WithDirectives);
-        buf.use_helper(Helper::WithModifiers);
-        buf.push("_withDirectives(node, [_withModifiers(handler), _withKeys(handler)])");
-
-        assert!(
-            buf.preamble().starts_with(
-                "const { withDirectives: _withDirectives, withModifiers: _withModifiers, withKeys: _withKeys"
-            )
-        );
-    }
-}
+mod tests;
